fixes to start, more restrictive genome

+++ jointPair.cpp
@@ -18,12 +18,12 @@ void JointPair::update() {
   _rightJoint->update();
 }
 
-void JointPair::setJointStats(double posAmp, double negAmp, double neutralAngle, double phase, double T) {
+void JointPair::setJointStats(float posAmp, float negAmp, float neutralAngle, float phase, float T) {
   _leftJoint->setJointStats(posAmp, negAmp, neutralAngle, phase, T);
   _rightJoint->setJointStats(posAmp, negAmp, neutralAngle, phase + 0.5, T);
 }
 
-void JointPair::prepareMoveToNeutralAngle(double T_ms) {
+void JointPair::prepareMoveToNeutralAngle(float T_ms) {
   _leftJoint->prepareMoveToNeutralAngle(T_ms);
   _rightJoint->prepareMoveToNeutralAngle(T_ms);
 }
